Refactoring for v3
