+++ src/MailChimp.php
@@ -1,0 +1,112 @@
+<?php
+
+namespace DrewM;
+
+/**
+ * Super-simple, minimum abstraction MailChimp API v3 wrapper
+ *
+ * Uses curl if available, falls back to file_get_contents and HTTP stream.
+ * This probably has more comments than code.
+ *
+ * @author Drew McLellan <drew.mclellan@gmail.com>
+ * @version 2.0
+ */
+class MailChimp
+{
+    private $api_key;
+    private $api_endpoint = 'https://<dc>.api.mailchimp.com/3.0';
+    private $verify_ssl   = true;
+
+    /**
+     * Create a new instance
+     * @param string $api_key Your MailChimp API key
+     */
+    public function __construct($api_key)
+    {
+        $this->api_key = $api_key;
+        list(, $datacentre) = explode('-', $this->api_key);
+        $this->api_endpoint = str_replace('<dc>', $datacentre, $this->api_endpoint);
+    }
+	
+	/**
+     * Validates MailChimp API Key
+     */
+    public function validateApiKey()
+    {
+        $request = $this->call('helper/ping');
+		return !empty($request);
+    }
+
+    /**
+     * Call an API method. Every request needs the API key, so that is added automatically -- you don't need to pass it in.
+     * @param  string $$http_verb   The HTTP verb to use: get, post, put, patch, delete
+     * @param  string $method       The API method to call, e.g. 'lists/list'
+     * @param  array  $args         An array of arguments to pass to the method. Will be json-encoded for you.
+     * @return array                Associative array of json decoded API response.
+     */
+    public function call($http_verb, $method, $args = array(), $timeout = 10)
+    {
+        $http_verb = strtolower($http_verb);
+        $verbs     = array('get', 'post', 'put', 'patch', 'delete');
+
+        if (!in_array($http_verb, $verbs)) {
+            throw new \Exception("Invalid HTTP verb. Must be one of ".implode(', ', $verbs));
+        }
+
+        return $this->makeRequest($http_verb, $method, $args, $timeout);
+    }
+
+    /**
+     * Performs the underlying HTTP request. Not very exciting
+     * @param  string $$http_verb   The HTTP verb to use: get, post, put, patch, delete
+     * @param  string $method       The API method to be called
+     * @param  array  $args         Assoc array of parameters to be passed
+     * @return array                Assoc array of decoded result
+     */
+    private function makeRequest($http_verb, $method, $args=array(), $timeout=10)
+    {
+        $args['apikey'] = $this->api_key;
+
+        $url = $this->api_endpoint.'/'.$method.'.json';
+        $json_data = json_encode($args);
+
+        if (function_exists('curl_init') && function_exists('curl_setopt')) {
+            $ch = curl_init();
+            curl_setopt($ch, CURLOPT_URL, $url);
+            curl_setopt($ch, CURLOPT_HTTPAUTH, CURLAUTH_BASIC); 
+            curl_setopt($ch, CURLOPT_USERPWD, 'drewm:'.$this->api_key); 
+            curl_setopt($ch, CURLOPT_HTTPHEADER, array('Accept: application/vnd.api+json',
+                                                        'Content-Type: application/vnd.api+json'));
+            curl_setopt($ch, CURLOPT_USERAGENT, 'DrewM/MailChimp-API/3.0 (github.com/drewm/mailchimp-api)');
+            curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+            curl_setopt($ch, CURLOPT_TIMEOUT, $timeout);
+            curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, $this->verify_ssl);
+            curl_setopt($ch, CURLOPT_HTTP_VERSION, CURL_HTTP_VERSION_1_0);
+
+
+            switch($http_verb) {
+                case 'post':
+                    curl_setopt($ch, CURLOPT_POST, true);
+                    curl_setopt($ch, CURLOPT_POSTFIELDS, $json_data); 
+                    break;
+
+                case 'get':
+                    $query = http_build_query($args);
+                    curl_setopt($ch, CURLOPT_URL, $url.'?'.$query);
+                    break;
+
+                case 'delete':
+                    curl_setopt($ch, CURLOPT_CUSTOMREQUEST, 'DELETE');
+                    break;
+            }
+
+
+            $result = curl_exec($ch);
+            curl_close($ch);
+        } else {
+            throw new \Exception("cURL support is required, but can't be found.");
+        }
+
+        return $result ? json_decode($result, true) : false;
+    }
+}
